+++ packages/gallery/src/settings/options/textBoxWidth.js
@@ -1,0 +1,12 @@
+import {GALLERY_CONSTS, INPUT_TYPES} from '../utils/constants';
+
+export default {
+  title: 'Text Box Width',
+  description: `Set the text box width when on the right side or on the left side.`,
+  isRelevant: (styleParams) => (styleParams.isVertical &&
+    styleParams.groupSize === 1 &&
+    !styleParams.oneRow) &&
+    (styleParams.allowTitle || styleParams.allowDescription),
+  type: INPUT_TYPES.NUMBER,
+  default: 200,
+}