+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 1.7.15'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 1.7.22'); } catch (e) {}